pdf-embedder/backend/requirements.txt
Flask==3.0.3
pypdf==4.3.1
reportlab==4.2.2
Pillow==10.4.0